split god class into several files
