+++ source/MechEngineMod.csproj
@@ -66,6 +66,24 @@
   <ItemGroup>
     <Compile Include="Control.cs" />
     <Compile Include="AssemblyInfo.cs" />
+    <Compile Include="EndoSteelMechInfoWidgetCalculateTonnagePatch.cs" />
+    <Compile Include="EndoSteelMechStatisticsRulesPatch.cs" />
+    <Compile Include="EngineMechComponentDamagePatch.cs" />
+    <Compile Include="EndoSteelMechValidationRulesPatch.cs" />
+    <Compile Include="EngineMechInitEffectStatsPatch.cs" />
+    <Compile Include="EngineMechLabInventoryWidgetPatch.cs" />
+    <Compile Include="EngineMechStatisticsRulesStatPatch.cs" />
+    <Compile Include="EngineSimGameStateWorkOrderPatch.cs" />
+    <Compile Include="GyroMechLabLocationWidgetPatch.cs" />
+    <Compile Include="GyroUpgradeDefPatch.cs" />
+    <Compile Include="InitialTonnageChassisDefPatch.cs" />
+    <Compile Include="Engine.cs" />
+    <Compile Include="EngineCalculator.cs" />
+    <Compile Include="MechEngineModSettings.cs" />
+    <Compile Include="MechLabLocationWidgetAdapter.cs" />
+    <Compile Include="EngineMechLabLocationWidgetPatch.cs" />
+    <Compile Include="EngineMechValidationRulesPatch.cs" />
+    <Compile Include="MechLabMechInfoWidgetAdapter.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\mod.json" />
